Register the mknote package with the Exif pkg at init
The caller can now cause the MakerNote section of an Exif file to be
parsed into the exif.Exif structure simply by importing the mknote
package.  This leaves the usage pattern of the exif library unchanged;
simply importing the mknote package to get access to its field names
is enough to cause those fields to be parsed from Exif files.